Pulled WatcherGroup out into its own file and decoupled it from HotReloadableViewController
