+++ BeatSaberMarkupLanguage/BeatSaberMarkupLanguage.csproj
@@ -125,11 +125,13 @@
     <Compile Include="Animations\GIF\GIFUnityDecoder.cs" />
     <Compile Include="Animations\AnimationInfo.cs" />
     <Compile Include="Animations\LockBitmap.cs" />
+    <Compile Include="Attributes\HotReloadAttribute.cs" />
     <Compile Include="Attributes\UIAction.cs" />
     <Compile Include="Attributes\UIComponent.cs" />
     <Compile Include="Attributes\UIObject.cs" />
     <Compile Include="Attributes\UIParams.cs" />
     <Compile Include="Attributes\UIValue.cs" />
+    <Compile Include="Attributes\ViewDefinitionAttribute.cs" />
     <Compile Include="BeatSaberUI.cs" />
     <Compile Include="Components\BSMLScrollView.cs" />
     <Compile Include="Components\BSMLTableView.cs" />
@@ -276,10 +278,12 @@
     <Compile Include="TypeHandlers\TextSegmentedControlHandler.cs" />
     <Compile Include="TypeHandlers\TypeHandler.cs" />
     <Compile Include="Utilities.cs" />
+    <Compile Include="ViewControllers\BSMLAutomaticViewController.cs" />
     <Compile Include="ViewControllers\BSMLResourceViewController.cs" />
     <Compile Include="ViewControllers\BSMLViewController.cs" />
     <Compile Include="ViewControllers\HotReloadableViewController.cs" />
     <Compile Include="ViewControllers\TestViewController.cs" />
+    <Compile Include="ViewControllers\WatcherGroup.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Views\test.bsml" />
